requirements.txt
asgiref==3.3.4
astroid==2.5.1
autopep8==1.5.6
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
cycler==0.10.0
defusedxml==0.7.1
dj-rest-auth==2.1.3
Django==3.2
django-allauth==0.44.0
django-phone-field==1.8.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
isort==5.7.0
kiwisolver==1.3.1
lazy-object-proxy==1.5.2
matplotlib==3.4.2
mccabe==0.6.1
mysqlclient==2.0.3
numpy==1.20.3
oauthlib==3.1.0
pandas==1.2.4
Pillow==8.2.0
protobuf==3.15.8
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.0.1
pylint==2.7.2
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
scipy==1.6.3
seaborn==0.11.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.3
wrapt==1.12.1
